Extract shared normalizeRepositoryURL helper
Pull the URL normalization out of CheckoutLocator into a module-internal
free function so the upcoming direct-dependency filter can reuse the same
SSH/HTTPS/.git/case normalization for matching.
